requirements.txt
Flask>=2.2.2
pdf2image>=1.14.0
Pillow==10.3.0
requests==2.31.0
opencv-python==4.9.0.80
numpy~=1.26.4